add API route
requires `FEATURE_FLAG_API_ENABLED` env variable to
be set to true in order to work.
