+++ .env.example
@@ -2,4 +2,4 @@ DATABASE_URL=
 SECRET=
 UPSTASH_REDIS_REST_URL=
 UPSTASH_REDIS_REST_TOKEN=
-API_ENABLED=true
+FEATURE_FLAG_API_ENABLED=true
