+++ app/routes/$season/index.tsx
@@ -1,6 +1,5 @@
 import { type Regions } from "@prisma/client";
-import { type LoaderArgs, type TypedResponse } from "@remix-run/node";
-import { json } from "@remix-run/node";
+import { json, type LoaderArgs, type TypedResponse } from "@remix-run/node";
 import { useLoaderData, useNavigation } from "@remix-run/react";
 import { type HeadersFunction, redirect } from "@remix-run/server-runtime";
 import Highcharts, {
@@ -15,29 +14,14 @@ import { Fragment, useEffect, useRef } from "react";
 
 import { getAffixIconUrl, getAffixName } from "~/affixes";
 import {
-  calculateExtrapolation,
-  calculateXAxisPlotLines,
-  calculateZoom,
-  determineExpirationTimestamp,
-  determineExtrapolationEnd,
   determineOverlaysToDisplayFromCookies,
   determineOverlaysToDisplayFromSearchParams,
   determineRegionsToDisplayFromCookies,
   determineRegionsToDisplayFromSearchParams,
-  loadDataForRegion,
 } from "~/load.server";
-import {
-  type EnhancedSeason,
-  findSeasonByName,
-  hasSeasonEndedForAllRegions,
-} from "~/seasons";
-import {
-  calculateFactionDiffForWeek,
-  isNotNull,
-  orderedRegionsBySize,
-  overlays as defaultOverlays,
-  searchParamSeparator,
-} from "~/utils";
+import { getEnhancedSeason } from "~/models/season.server";
+import { type EnhancedSeason, findSeasonByName } from "~/seasons";
+import { calculateFactionDiffForWeek, searchParamSeparator } from "~/utils";
 
 const factionColors = {
   alliance: "#60a5fa",
@@ -108,16 +92,6 @@ export const loader = async ({
     });
   }
 
-  const headers: HeadersInit = {};
-
-  if (hasSeasonEndedForAllRegions(season.slug)) {
-    const thirtyDays = 30 * 24 * 60 * 60;
-    headers[
-      cacheControl
-    ] = `public, max-age=${thirtyDays}, s-maxage=${thirtyDays}, immutable`;
-  }
-
-  const extrapolationEnd = determineExtrapolationEnd(request);
   const searchParamOverlays =
     determineOverlaysToDisplayFromSearchParams(request);
   const searchParamRegions = determineRegionsToDisplayFromSearchParams(request);
@@ -143,113 +117,15 @@ export const loader = async ({
     return redirect(`/${season.slug}?${params.toString()}`, 307);
   }
 
-  const regions = searchParamRegions ?? orderedRegionsBySize;
-  const overlays = searchParamOverlays ?? defaultOverlays;
+  const regions = searchParamRegions;
+  const overlays = searchParamOverlays;
 
-  const enhancedSeason: EnhancedSeason = {
-    ...season,
-    regionsToDisplay: regions,
-    overlaysToDisplay: [...overlays],
-    dataByRegion: {
-      eu: [],
-      us: [],
-      kr: [],
-      tw: [],
-    },
-    extrapolation: {
-      eu: null,
-      kr: null,
-      tw: null,
-      us: null,
-    },
-    initialZoom: {
-      eu: null,
-      kr: null,
-      tw: null,
-      us: null,
-    },
-    xAxisPlotLines: {
-      eu: [],
-      us: [],
-      kr: [],
-      tw: [],
-    },
-  };
-
-  const now = Date.now();
-
-  await Promise.all(
-    Object.values(regions).map(async (region) => {
-      const data = await loadDataForRegion(region, season);
-      enhancedSeason.dataByRegion[region] = data;
-
-      if (data.length === 0) {
-        return;
-      }
-
-      const extrapolation = calculateExtrapolation(
-        season,
-        region,
-        data,
-        extrapolationEnd
-      );
-
-      enhancedSeason.xAxisPlotLines[region] = calculateXAxisPlotLines(
-        season,
-        region,
-        data,
-        extrapolation,
-        overlays
-      );
-
-      const seasonEnding = season.endDates[region];
-
-      // season ended, no need for zoomies or extrapolation
-      if (seasonEnding && seasonEnding <= now) {
-        return;
-      }
-
-      enhancedSeason.extrapolation[region] = extrapolation;
-
-      enhancedSeason.initialZoom[region] = calculateZoom(
-        season,
-        region,
-        data,
-        extrapolation
-      );
-    })
-  );
-
-  const mostRecentDataset = Object.values(enhancedSeason.dataByRegion)
-    .flat()
-    .reduce((acc, dataset) => (acc > dataset.ts ? acc : dataset.ts), 0);
-
-  headers[lastModified] = new Date(mostRecentDataset).toUTCString();
-
-  const shortestExpiry = regions
-    .map((region) =>
-      determineExpirationTimestamp(
-        season,
-        region,
-        enhancedSeason.dataByRegion[region]
-      )
-    )
-    .reduce(
-      (acc, expiry) => (acc > expiry ? expiry : acc),
-      Number.POSITIVE_INFINITY
-    );
-
-  headers[expires] = new Date(shortestExpiry * 1000 + Date.now()).toUTCString();
-  headers[eTag] = [
-    season.slug,
-    mostRecentDataset,
-    extrapolationEnd,
-    ...regions,
-    ...overlays,
-  ]
-    .filter(isNotNull)
-    .sort((a, b) => (a > b ? 1 : -1))
-    .join("-");
+  const { season: enhancedSeason, headers } = await getEnhancedSeason({
+    request,
+    regions,
+    overlays,
+    season,
+  });
 
   return json(enhancedSeason, { headers });
 };
